==> R/01_01_lectura_datos_originales_rds.R <==

bahamas_leaks_edges = read_rds(path = "./salidas_datos/originales/bahamas_leaks_edges.rds")
bahamas_leaks_nodes_address = read_rds(path = "./salidas_datos/originales/bahamas_leaks_nodes_address.rds")
bahamas_leaks_nodes_entity = read_rds(path = "./salidas_datos/originales/bahamas_leaks_nodes_entity.rds")
bahamas_leaks_nodes_intermediary = read_rds(path = "./salidas_datos/originales/bahamas_leaks_nodes_intermediary.rds")
bahamas_leaks_nodes_officer = read_rds(path = "./salidas_datos/originales/bahamas_leaks_nodes_officer.rds")

offshore_leaks_edges = read_rds(path = "./salidas_datos/originales/offshore_leaks_edges.rds")
offshore_leaks_nodes_address = read_rds(path = "./salidas_datos/originales/offshore_leaks_nodes_address.rds")
offshore_leaks_nodes_entity = read_rds(path = "./salidas_datos/originales/offshore_leaks_nodes_entity.rds")
offshore_leaks_nodes_intermediary = read_rds(path = "./salidas_datos/originales/offshore_leaks_nodes_intermediary.rds")
offshore_leaks_nodes_officer = read_rds(path = "./salidas_datos/originales/offshore_leaks_nodes_officer.rds")

panama_papers_edges = read_rds(path = "./salidas_datos/originales/panama_papers_edges.rds")
panama_papers_nodes_address = read_rds(path = "./salidas_datos/originales/panama_papers_nodes_address.rds")
panama_papers_nodes_entity = read_rds(path = "./salidas_datos/originales/panama_papers_nodes_entity.rds")
panama_papers_nodes_intermediary = read_rds(path = "./salidas_datos/originales/panama_papers_nodes_intermediary.rds")
panama_papers_nodes_officer = read_rds(path = "./salidas_datos/originales/panama_papers_nodes_officer.rds")

paradise_papers_edges = read_rds(path = "./salidas_datos/originales/paradise_papers_edges.rds")
paradise_papers_nodes_address = read_rds(path = "./salidas_datos/originales/paradise_papers_nodes_address.rds")
paradise_papers_nodes_entity = read_rds(path = "./salidas_datos/originales/paradise_papers_nodes_entity.rds")
paradise_papers_nodes_intermediary = read_rds(path = "./salidas_datos/originales/paradise_papers_nodes_intermediary.rds")
paradise_papers_nodes_officer = read_rds(path = "./salidas_datos/originales/paradise_papers_nodes_officer.rds")
